requirements.txt
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
distlib==0.3.0
Django==3.0.5
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.5.0
django-filter==2.4.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djongo==1.3.3
djoser==2.1.0
filelock==3.0.12
freeze==3.0
idna==2.10
importlib-metadata==1.6.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==5.0.0
mysql-connector==2.2.9
numpy==1.17.2
oauthlib==3.1.0
Pillow==7.1.1
Pillow-PIL==0.1.dev0
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.1
PyMySQL==0.9.3
pytest==3.8.0
python3-openid==3.2.0
pytz==2020.1
request==1.0.114
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.23
zipp==3.1.0
